requirements.txt
pillow
djangorestframework
django-cors-headers
